Ped Randomizer: Player randomizer improvements + lots of crash fixes
* Cutscene Randomizer will no longer randomize ped models because it
has a million issues with ped registering. Ped Randomizer does it
pretty well including player randomization. It'll be renamed to
Cutscene Prop Randomizer in the future

* Cops are now random with Ped Randomizer and cops can spawn as normal
peds (and won't bust you).

* Fish/Rat related skeleton crashes have been fixed for the most
part. There might be few edge cases that need to be discovered for it
to be fully functional though. This lays ground for future work on ped
randomizer and allowing animals to be used as normal peds in missions.